# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: FortKnoxx Security Intelligence Platform
  description: 'Unified security scanning API: SAST, DAST, SCA, IaC, secrets, container, and AI-powered scanners (zero-day
    GNN, business logic, LLM adversarial, auth).'
  version: 1.0.0
paths:
...
  /api/stats/{repo_id}:
    get:
      tags:
      - stats
      summary: Get Repository Stats
      operationId: get_repository_stats_api_stats__repo_id__get
      parameters:
      - name: repo_id
        in: path
        required: true
        schema:
          type: string
          title: Repo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Repository Stats Api Stats  Repo Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/settings/api-keys:
    get:
      tags:
      - settings
      summary: Get Api Keys Values
      description: 'Return masked key values for the admin UI.


        WARNING: only the first 8 characters of each key are returned;

        full plaintext is never sent to the frontend.'
      operationId: get_api_keys_values_api_settings_api_keys_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Get Api Keys Values Api Settings Api Keys Get
    post:
      tags:
      - settings
      summary: Update Api Keys
      description: 'Encrypted store of LLM provider keys + scanner tokens.


        Pass an empty string to delete a single key.'
      operationId: update_api_keys_api_settings_api_keys_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAPIKeysRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Update Api Keys Api Settings Api Keys Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/settings/scanners:
    get:
      tags:
      - settings
      summary: Get Installed Scanners
      description: "Which scanner binaries are reachable on PATH at this moment.\n\nIndependent of the persisted enable/disable\
        \ state \u2014 useful for\nthe UI to show a red badge next to a scanner that is enabled in\nsettings but missing from\
        \ the host. Persisted config lives at\n``GET /settings/scanners/config``."
      operationId: get_installed_scanners_api_settings_scanners_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  additionalProperties: true
                  type: object
                type: object
                title: Response Get Installed Scanners Api Settings Scanners Get
    put:
      tags:
      - settings
      summary: Update Scanner Config
      operationId: update_scanner_config_api_settings_scanners_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateScannerSettingsRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Update Scanner Config Api Settings Scanners Put
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/settings/ai-scanners:
    get:
      tags:
      - settings
      summary: Get Ai Scanner Settings
      operationId: get_ai_scanner_settings_api_settings_ai_scanners_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIScannerSettings'
    post:
      tags:
      - settings
      summary: Update Ai Scanner Settings
      operationId: update_ai_scanner_settings_api_settings_ai_scanners_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAIScannerSettingsRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Update Ai Scanner Settings Api Settings Ai Scanners Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AIScannerSettings:
      properties:
        enable_zero_day_detector:
          type: boolean
          title: Enable Zero Day Detector
          description: Enable ML-based zero-day detection
          default: true
        enable_business_logic_scanner:
          type: boolean
          title: Enable Business Logic Scanner
          description: Enable business logic flaw detection
          default: true
        enable_llm_security_scanner:
          type: boolean
          title: Enable Llm Security Scanner
          description: Enable LLM prompt injection testing
          default: true
        enable_auth_scanner:
          type: boolean
          title: Enable Auth Scanner
          description: Enable authentication vulnerability scanner
          default: true
      type: object
      title: AIScannerSettings
      description: 'DEPRECATED: Use ScannerSettings instead. Kept for backwards compatibility.'
      example:
        enable_auth_scanner: false
        enable_business_logic_scanner: true
        enable_llm_security_scanner: true
        enable_zero_day_detector: true
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UpdateAIScannerSettingsRequest:
      properties:
        enable_zero_day_detector:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Zero Day Detector
        enable_business_logic_scanner:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Business Logic Scanner
        enable_llm_security_scanner:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Llm Security Scanner
        enable_auth_scanner:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Auth Scanner
      type: object
      title: UpdateAIScannerSettingsRequest
      description: 'DEPRECATED: Use UpdateScannerSettingsRequest instead. Kept for backwards compatibility.'
    UpdateAPIKeysRequest:
      properties:
        openai_api_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Openai Api Key
        anthropic_api_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Anthropic Api Key
        gemini_api_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Gemini Api Key
        github_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Github Token
        snyk_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Snyk Token
      type: object
      title: UpdateAPIKeysRequest
      description: Request model for updating API keys
      example:
        anthropic_api_key: sk-ant-...
        gemini_api_key: AI...
        github_token: ghp_...
        openai_api_key: sk-...
        snyk_token: '...'
    UpdateScannerSettingsRequest:
      properties:
        enable_semgrep:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Semgrep
        enable_bandit:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Bandit
        enable_gitleaks:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Gitleaks
        enable_trufflehog:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Trufflehog
        enable_trivy:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Trivy
        enable_grype:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Grype
        enable_checkov:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Checkov
        enable_eslint:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Eslint
        enable_pylint:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Pylint
        enable_flake8:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Flake8
        enable_radon:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Radon
        enable_shellcheck:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Shellcheck
        enable_hadolint:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Hadolint
        enable_sqlfluff:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Sqlfluff
        enable_pydeps:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Pydeps
        enable_pip_audit:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Pip Audit
        enable_npm_audit:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Npm Audit
        enable_syft:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Syft
        enable_nuclei:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Nuclei
        enable_codeql:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Codeql
        enable_snyk:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Snyk
        enable_gosec:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Gosec
        enable_spotbugs:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Spotbugs
        enable_pyre:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Pyre
        enable_horusec:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Horusec
        enable_zap:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Zap
        enable_zap_dast:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Zap Dast
        enable_api_fuzzer:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Api Fuzzer
        enable_zero_day_detector:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Zero Day Detector
        enable_business_logic_scanner:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Business Logic Scanner
        enable_llm_security_scanner:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Llm Security Scanner
        enable_auth_scanner:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Auth Scanner
      type: object
      title: UpdateScannerSettingsRequest
      description: Request model for updating scanner settings - any scanner can be enabled/disabled
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...